requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
PyMySQL==1.1.0
pyodbc==5.1.0
python-dotenv==1.0.0
Werkzeug==2.3.7
bcrypt==4.0.1
jwt==1.3.1
requests==2.31.0
